Create the folder if it doesn't exist already.

+++ test/spock-functional-test/src/test/groovy/framework/ReposeConfigurationProvider.groovy
@@ -94,7 +94,12 @@ class ReposeConfigurationProvider {
     }
 
     public void cleanConfigDirectory() {
-        FileUtils.cleanDirectory(reposeConfigDir)
+        if (reposeConfigDir.exists()) {
+            FileUtils.cleanDirectory(reposeConfigDir)
+        } else {
+            reposeConfigDir.mkdirs()
+        }
+
     }
 
 }
